+++ plugins/org.eclipse.osee.framework.skynet.core/src/org/eclipse/osee/framework/skynet/core/transaction/TransactionManager.java
@@ -181,7 +181,7 @@ public final class TransactionManager {
       String comment = stmt.getString("osee_comment");
       Date timestamp = stmt.getTimestamp("time");
       UserId authorArtId = UserId.valueOf(stmt.getLong("author"));
-      Integer commitArtId = stmt.getInt("commit_art_id");
+      Long commitArtId = stmt.getLong("commit_art_id");
       Long buildId = stmt.getLong("build_id");
       TransactionDetailsType txType = TransactionDetailsType.valueOf(stmt.getInt("tx_type"));
       return new TransactionRecord(transactionNumber, branch, comment, timestamp, authorArtId, commitArtId, txType,
