+++ plugins/org.eclipse.osee.framework.ui.branch.graph/src/org/eclipse/osee/framework/ui/branch/graph/model/TxData.java
@@ -28,11 +28,11 @@ public class TxData {
    private final Date timeStamp;
    private final String comment;
    private final TransactionDetailsType txType;
-   private final int commitArtId;
+   private final Long commitArtId;
    private final BranchId branch;
    private final Long txId;
 
-   public TxData(BranchId branch, UserId authorId, Date timeStamp, String comment, TransactionDetailsType txType, int commitArtId, long txId) {
+   public TxData(BranchId branch, UserId authorId, Date timeStamp, String comment, TransactionDetailsType txType, Long commitArtId, long txId) {
       this.authorId = authorId;
       this.timeStamp = timeStamp;
       this.comment = comment;
@@ -82,7 +82,7 @@ public class TxData {
    /**
     * @return the commitArtId
     */
-   public int getCommitArtId() {
+   public Long getCommitArtId() {
       return commitArtId;
    }
